Revert "[Op] Optimize list of GPU SegmentReductionOps. (#783)"
This reverts commit ff9d339827994fb6ef9e3fcd513d1ce16502b1cd.
